+++ test/migration_helpers_test.rb
@@ -16,6 +16,22 @@ class AddNullConstraintSafely < TestMigration
   end
 end
 
+class AddColumnSafely < TestMigration
+  disable_ddl_transaction!
+
+  def change
+    add_column_safely :users, :balance, :integer, default: 10, null: false
+  end
+end
+
+class BackfillColumnSafely < TestMigration
+  disable_ddl_transaction!
+
+  def up
+    backfill_column_safely :users, :city, "San Francisco"
+  end
+end
+
 class MigrationHelpersTest < Minitest::Test
   def setup
     skip unless ENV["HELPERS"]
@@ -56,6 +72,42 @@ class MigrationHelpersTest < Minitest::Test
 
     migrate(AddNullConstraintSafely)
     migrate(AddNullConstraintSafely, direction: :down)
+  end
+
+  def test_add_column_safely_raises_inside_transaction
+    skip unless postgresql?
+    error = assert_raises(StrongMigrations::Error) { migrate(AddColumnSafely, transaction: true) }
+    assert_match "Cannot run `add_column_safely` inside a transaction", error.message
+  end
+
+  def test_add_column_safely
+    User.reset_column_information
+    migrate(AddColumnSafely)
+
+    column = User.columns.find { |c| c.name == "balance" }
+    assert_equal :integer, column.type
+    assert_equal "10", column.default
+    assert_equal false, column.null
+  ensure
+    migrate(AddColumnSafely, direction: :down)
+    User.reset_column_information
+  end
+
+  def test_backfill_column_safely_raises_inside_transaction
+    error = assert_raises(StrongMigrations::Error) { migrate(BackfillColumnSafely, transaction: true) }
+    assert_match "Cannot run `backfill_column_safely` inside a transaction", error.message
+  end
+
+  def test_backfill_column_safely
+    user1 = User.create(name: "John", city: "Los Angeles")
+    user2 = User.create(name: "Jane", city: nil)
+    migrate(BackfillColumnSafely)
+
+    assert_equal "Los Angeles", user1.reload.city
+    assert_equal "San Francisco", user2.reload.city
+  ensure
+    User.delete_all
+    User.reset_column_information
   end
 
   private